+++ src/console.hpp
@@ -77,15 +77,15 @@ class console_t
 
   void send(interrupt_type_t flag);
 
-  uint32_t read_byte(uint32_t address);
+  uint8_t read_byte(uint32_t address);
 
-  uint32_t read_half(uint32_t address);
+  uint16_t read_half(uint32_t address);
 
   uint32_t read_word(uint32_t address);
 
-  void write_byte(uint32_t address, uint32_t data);
+  void write_byte(uint32_t address, uint8_t data);
 
-  void write_half(uint32_t address, uint32_t data);
+  void write_half(uint32_t address, uint16_t data);
 
   void write_word(uint32_t address, uint32_t data);
