+++ src/memory-component.cpp
@@ -26,12 +26,12 @@ void memory_component_t::log(const char *format, ...) {
   }
 }
 
-uint32_t memory_component_t::io_read_byte(uint32_t address) {
+uint8_t memory_component_t::io_read_byte(uint32_t address) {
   log("io_read_byte(0x%08x)", address);
   return 0;
 }
 
-uint32_t memory_component_t::io_read_half(uint32_t address) {
+uint16_t memory_component_t::io_read_half(uint32_t address) {
   log("io_read_half(0x%08x)", address);
   return 0;
 }
@@ -41,11 +41,11 @@ uint32_t memory_component_t::io_read_word(uint32_t address) {
   return 0;
 }
 
-void memory_component_t::io_write_byte(uint32_t address, uint32_t data) {
+void memory_component_t::io_write_byte(uint32_t address, uint8_t data) {
   log("io_write_byte(0x%08x, 0x%08x)", address, data);
 }
 
-void memory_component_t::io_write_half(uint32_t address, uint32_t data) {
+void memory_component_t::io_write_half(uint32_t address, uint16_t data) {
   log("io_write_half(0x%08x, 0x%08x)", address, data);
 }
