+++ src/spu/core.cpp
@@ -12,7 +12,7 @@ core_t::core_t(bool log_enabled)
   , sound_ram("sound-ram") {
 }
 
-uint32_t core_t::io_read_half(uint32_t address) {
+uint16_t core_t::io_read_half(uint32_t address) {
   if (address >= 0x1f801c00 && address <= 0x1f801d7f) {
     auto n = (address >> 4) & 31;
     auto m = (address >> 1) & 7;
@@ -91,7 +91,7 @@ uint32_t core_t::io_read_half(uint32_t address) {
   return memory_component_t::io_read_half(address);
 }
 
-void core_t::io_write_half(uint32_t address, uint32_t data) {
+void core_t::io_write_half(uint32_t address, uint16_t data) {
   if (address >= 0x1f801c00 && address <= 0x1f801d7f) {
     auto n = (address >> 4) & 31;
     auto m = (address >> 1) & 7;
